[#7] refactor: separate matching types into matching.ts
